requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
accelerate=0.23.0=pypi_0
aiohttp=3.8.6=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.6.0=pypi_0
antlr4-python3-runtime=4.8=pypi_0
asttokens=2.4.0=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
audioread=3.0.1=pypi_0
av=10.0.0=pypi_0
backcall=0.2.0=pypi_0
backoff=2.2.1=pypi_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
bidict=0.22.1=pypi_0
bitarray=2.8.2=pypi_0
blinker=1.6.3=pypi_0
bzip2=1.0.8=hd590300_5
ca-certificates=2023.11.17=hbcca054_0
cairo=1.18.0=h3faef2a_0
certifi=2023.7.22=pypi_0
cffi=1.16.0=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.3.0=pypi_0
click=8.1.7=pypi_0
cmake=3.27.6=pypi_0
colorama=0.4.6=pypi_0
coloredlogs=15.0.1=pypi_0
comm=0.1.4=pyhd8ed1ab_0
ctranslate2=3.21.0=pypi_0
cython=3.0.3=pypi_0
datasets=2.14.6=pypi_0
debugpy=1.6.7=py39h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
dill=0.3.7=pypi_0
distance=0.1.3=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
executing=2.0.0=pypi_0
expat=2.5.0=hcb278e6_1
fairseq=0.12.2=pypi_0
faster-whisper=0.9.0=pypi_0
ffmpeg-python=0.2.0=pypi_0
ffmpegio=0.9.0=pypi_0
ffmpegio-core=0.9.0=pypi_0
filelock=3.12.4=pypi_0
flask=3.0.0=pypi_0
flask-socketio=5.3.6=pypi_0
flatbuffers=23.5.26=pypi_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_1
fontconfig=2.14.2=h14ed4e7_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
freetype=2.12.1=h267a509_2
frozenlist=1.4.0=pypi_0
fsspec=2023.6.0=pypi_0
ftfy=6.1.1=pypi_0
future=0.18.3=pypi_0
g2p-en=2.1.0=pypi_0
gettext=0.21.1=h27087fc_0
h11=0.14.0=pypi_0
huggingface-hub=0.17.3=pypi_0
humanfriendly=10.0=pypi_0
hydra-core=1.0.7=pypi_0
icu=73.2=h59595ed_0
idna=3.4=pypi_0
importlib-metadata=6.8.0=pypi_0
inflect=7.0.0=pypi_0
ipykernel=6.26.0=pyhf8b6a83_0
ipython=8.16.1=pypi_0
itsdangerous=2.1.2=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.5.0=py39hf3d152e_0
lazy-loader=0.3=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libexpat=2.5.0=hcb278e6_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.2.0=h807b86a_3
libgirepository=1.78.1=h003a4f0_1
libglib=2.78.3=h783c2da_0
libiconv=1.17=hd590300_2
libpng=1.6.39=h753d276_0
librosa=0.10.1=pypi_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=13.2.0=h7e041cc_3
libuuid=2.38.1=h0b41bf4_0
libxcb=1.15=h0b41bf4_0
libzlib=1.2.13=hd590300_5
lit=17.0.2=pypi_0
llvm-openmp=17.0.6=h4dfa4b3_0
llvmlite=0.41.0=pypi_0
loguru=0.7.2=pypi_0
lxml=4.9.3=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
more-itertools=10.1.0=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.0.7=pypi_0
multidict=6.0.4=pypi_0
multiprocess=0.70.15=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
networkx=3.1=pypi_0
nltk=3.8.1=pypi_0
numba=0.58.0=pypi_0
numpy=1.25.2=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.2.140=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
omegaconf=2.0.6=pypi_0
onnxruntime=1.16.2=pypi_0
openai=0.28.1=pypi_0
openai-whisper=20230918=pypi_0
openssl=3.2.0=hd590300_1
optimum=1.13.2=pypi_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.1=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pcre2=10.42=hcad00b1_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pip=23.3.2=pypi_0
pixman=0.42.2=h59595ed_0
platformdirs=3.11.0=pyhd8ed1ab_0
playsound2=1.3.0=pypi_0
pluggy=1.3.0=pypi_0
pooch=1.8.0=pypi_0
portalocker=2.8.2=pypi_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
protobuf=4.24.4=pypi_0
psutil=5.9.5=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=13.0.0=pypi_0
pyaudio=0.2.13=pypi_0
pycairo=1.25.1=pypi_0
pycparser=2.21=pypi_0
pydantic=2.4.2=pypi_0
pydantic-core=2.10.1=pypi_0
pydub=0.25.1=pypi_0
pyee=11.0.1=pypi_0
pygments=2.16.1=pyhd8ed1ab_0
pygobject=3.46.0=py39hb25b1be_1
python=3.9.18=h955ad1f_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-engineio=4.8.0=pypi_0
python-socketio=5.10.0=pypi_0
python_abi=3.9=2_cp39
pyttsx3=2.90=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.0=py39h6a678d5_0
readline=8.2=h5eee18b_0
regex=2023.10.3=pypi_0
requests=2.31.0=pypi_0
resampy=0.4.2=pypi_0
sacrebleu=2.3.1=pypi_0
safetensors=0.4.0=pypi_0
samplerate=0.1.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.11.3=pypi_0
sentencepiece=0.1.99=pypi_0
setuptools=68.0.0=py39h06a4308_0
simple-websocket=1.0.0=pypi_0
six=1.16.0=pyh6c4a22f_0
sounddevice=0.4.6=pypi_0
soundfile=0.12.1=pypi_0
soxr=0.3.7=pypi_0
speechrecognition=3.10.0=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pypi_0
tabulate=0.9.0=pypi_0
tensorboardx=2.6.2.2=pypi_0
threadpoolctl=3.2.0=pypi_0
tiktoken=0.5.1=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.14.1=pypi_0
torch=2.1.0=pypi_0
torchaudio=2.1.0=pypi_0
tornado=6.1=py39hb9d737c_3
tqdm=4.66.1=pypi_0
traitlets=5.11.2=pypi_0
transformers=4.35.0=pypi_0
triton=2.1.0=pypi_0
typing-extensions=4.8.0=hd8ed1ab_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2023.3=pypi_0
urllib3=2.0.6=pypi_0
wcwidth=0.2.8=pypi_0
webrtcvad=2.0.10=pypi_0
websocket-client=1.7.0=pypi_0
werkzeug=3.0.0=pypi_0
wheel=0.41.2=py39h06a4308_0
wsproto=1.2.0=pypi_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.1.1=hd590300_0
xorg-libsm=1.2.4=h7391055_0
xorg-libx11=1.8.7=h8ee46fc_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxrender=0.9.11=hd590300_0
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h0b41bf4_1003
xorg-xproto=7.0.31=h7f98852_1007
xxhash=3.4.1=pypi_0
xz=5.4.2=h5eee18b_0
yarl=1.9.2=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.17.0=pypi_0
zlib=1.2.13=hd590300_5
zstd=1.5.5=hfc55251_0
